Update test suites to run in docker containers
The existing setup where it builds in travis is now failing due to some dependency i've not traced updating and not being compatible with the old version of php. This is expected as I was running the tests with `--ignore-platform-reqs` and tbh i'm a bit surprised its taken this long to fail in a way that forced me to refactor the process. Just 3 and a half years :)

We need old and new versions of magento to composer install for the existing test coverage, we cannot rely on the versions of php supplied in phpenv within native travis as they're not always available. So I've built an image that does what we need.

- To support this I've had to rename the test envs from `24` to `m24` so they could live in a docker compose file (as they can't just be numeric)
    - This also means the test files have had to move about a bit too.
- You can now run the same setup locally without too much faffing about, although it is much slower on my macbook than in travis (likely because of the volume)
- I ensured support for PHP 8.1 and latest version of magento as part of this, this meant bumping some dependencies like symfony/console and php-cs-fixer
    - I have the composer.json a little modified actually so it technically could still run older symfony console version, but thats not the version we've locked it to. This is because I have some tests that still do a 2.2.0 to -> 2.2.11, I need to verify all override types and tests are reflected in 2.4 series before I move these test cases about a little, see the TODO in `.travis.yml`