Add extra error handling to db schema
I can only test against magento core, we're now doing analysis on a lot of db_schema.xml from a lot of random modules.

I dont want this to break the tool in these cases, if we can't read the file for some reason, suppress it and move on.

We'll see the message in the `-vvv` output